리뷰 페이지 기능 구현 완료
- 리뷰 뷰모델 구현 완료
- 리뷰 페이지 완료
- 리뷰 작성 구현 완료
- 코드 리펙토링
- 파이어베이스 업로드 및 리스트 불러오기 완료

+++ lib/data/model/location.dart
@@ -2,8 +2,8 @@ class Location {
   String? title;
   String? category;
   String? roadAddress;
-  int? mapx;
-  int? mapy;
+  double? mapx;
+  double? mapy;
 
   Location({this.title, this.category, this.roadAddress, this.mapx, this.mapy});
 
@@ -11,7 +11,7 @@ class Location {
     title = json["title"];
     category = json["category"];
     roadAddress = json["roadAddress"];
-    mapx = int.tryParse(json["mapx"]) ?? -1;
-    mapy = int.tryParse(json["mapy"]) ?? -1;
+    mapx = double.tryParse(json["mapx"]) ?? -1;
+    mapy = double.tryParse(json["mapy"]) ?? -1;
   }
 }
